inference/neuralJamAI/pyproject.toml
[project]
name = "neuraljamai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.7",
    "langchain>=0.3.15",
    "langchain-mistralai>=0.2.4",
    "langgraph>=0.2.67",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
